backend/ai_services/ml_service/requirements.txt
# Logging
loguru>=0.7.0

# Core ML & REMOSTAR
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
protobuf<4.0.0

# API framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# HTTP & async
requests>=2.31.0
httpx>=0.24.0

# Data processing
pandas>=2.0.0
polars>=0.19.0
openpyxl>=3.1.0
pyyaml>=6.0

# Neo4j integration
neo4j>=5.0.0
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0